+++ ppytools/csvhelper3.py
@@ -1,41 +1,35 @@
 # -*- -*- coding: utf-8 -*- -*-
 # __author__ = 'elkan1788@gmail.com'
 
-from ppytools.lang.timerhelper import timeMeter
+from ppytools.lang.timerhelper import timemeter
 
-import codecs
 import csv
 import logging
 
 logger = logging.getLogger(__name__)
 
-BOM_FORMAT = codecs.BOM_UTF8
 
-@timeMeter()
+@timemeter()
 def write(path, head, data):
     try:
-        with open(unicode(path, 'UTF-8'), 'wb') as csv_file:
-            csv_file.write(BOM_FORMAT)
+        with open(path, 'w', newline='', encoding='UTF-8-sig') as csv_file:
             writer = csv.writer(csv_file)
-
             if head is not None:
                 writer.writerow(head)
-
-            for row in data:
-                writer.writerow(row)
-
+            writer.writerows(data)
             logger.info('Write a CSV file successful. --> %s', path)
-    except Exception, e:
+    except Exception as e:
         raise Exception("Write a CSV file failed!!! --> %s, Case: %s" % (path, str(e)))
 
-@timeMeter()
+
+@timemeter()
 def getIdNameDict(path):
     id_name_dict = dict()
     try:
-        with open(path, 'rb') as tmp:
+        with open(path, 'r', encoding='UTF-8-sig') as tmp:
             reader = csv.reader(tmp)
             for line in reader:
                 id_name_dict[line[0]] = line[1]
-    except Exception, e:
+    except Exception as e:
         raise Exception("Read CSV file was failed!!! --> %s Case: %s" % (path, str(e)))
     return id_name_dict