+++ test/suites_start.py
@@ -1,29 +1,32 @@
 # -*- coding: utf-8 -*-
 # __author__ = 'elkan1788@gmail.com'
+
 from unittest import TestSuite
 
-from HTMLTestRunner import HTMLTestRunner
+import HTMLTestRunner
 import logging
 import os
 import unittest
+import sys
 
-logging.basicConfig(level=logging.DEBUG, encode='UTF-8', encoding='UTF-8',
-                    format='%(asctime)s [%(levelname)s] {%(name)-10s} - %(message)s')
+logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(levelname)s] {%(name)-10s} - %(message)s')
 
+logger = logging.getLogger(__name__)
 
 if __name__ == '__main__':
 
+    logger.info('Python ENV: %s', sys.version_info)
+
     suite = TestSuite()
-    suites = unittest.defaultTestLoader.discover('./')
-    for all_test_suite in suites:
-        for test_suite in all_test_suite:
-            suite.addTests(test_suite)
+    suites = unittest.defaultTestLoader.discover('./', pattern='test_*')
+    for test_suite in suites:
+        suite.addTests(test_suite)
 
     work_dir = os.path.dirname(os.path.realpath(__file__))
     report_html = os.path.join(work_dir, 'reports/ppytools_test_report.html')
 
     with open(report_html, 'wb') as html:
-        runner = HTMLTestRunner(stream=html, title='PPyTools Test Report')
+        runner = HTMLTestRunner.HTMLTestRunner(stream=html, title='PPyTools Test Report')
         runner.run(suite)