+++ test/test_emailclient.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
 # __author__ = 'elkan1788@gmail.com'
 
-from ppytools.cfgreader import ConfReader
+
 from ppytools.emailclient import EmailClient
 
 import os
 import logging
 import unittest
+import sys
+
+if sys.version_info[0] < 3:
+    from ppytools.cfgreader2 import ConfReader2 as ConfReader
+else:
+    from ppytools.cfgreader3 import ConfReader3 as ConfReader
 
 logger = logging.getLogger(__name__)