+++ HISTORY.rst
@@ -4,6 +4,10 @@ History
 .. image:: https://cdn.rawgit.com/sindresorhus/awesome/d7305f38d29fed78fa85652e3a63e154dd8e8829/media/badge.svg
     :alt: Awesome Python
 
+1.2.0(05/06/2018)
+-----------------
+- Update environment to support python2.7+
+
 1.1.0(28/02/2018)
 -----------------
 
@@ -22,7 +26,7 @@ Features
 -----------------
 
 - Fix release bugs
-- Finish components: configure, compress, csv, email, strings, timeMeter, hive, ip2location
+- Finish components: configure, compress, csv, email, strings, timemeter, hive, ip2location
 
 Features
 *********